+++ SwiftMessages/UIViewController+Utils.swift
@@ -21,7 +21,7 @@ extension UIViewController {
         } else if case .bottom? = topBottomStyle, let tabBarController = sm_selectTabBarControllerTopDown() {
             return tabBarController
         }
-        return WindowViewController(windowLevel: self.view.window?.windowLevel ?? UIWindowLevelNormal, config: config)
+        return WindowViewController.newInstance(windowLevel: self.view.window?.windowLevel, config: config)
     }
     
     fileprivate func sm_selectNavigationControllerTopDown() -> UINavigationController? {
@@ -80,7 +80,7 @@ extension UIViewController {
             if let parent = self.parent {
                 return parent.sm_selectPresentationContextBottomUp(config)
             } else {
-                return WindowViewController(windowLevel: self.view.window?.windowLevel ?? UIWindowLevelNormal, config: config)
+                return WindowViewController.newInstance(windowLevel: self.view.window?.windowLevel, config: config)
             }
         }
         return self