+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## 4.1.2
 
+### Features
+* #164 Added an optional `windowViewController` property to `SwiftMessages.Config` for supplying a custom subclass of `WindowViewController`.
+
 ### Bug Fixes
 * Custom presentation styles using `TopBottomAnimation` now display properly under top and bottom bars.